Allow cropping from full-sized images, and random scale augmentation by resizing an image between a min_size and max_size before crop.
As a side effect of allowing the use of full images, we reach a limit of protocol buffers messages - can no longer always use ParseFromString to parse an image from a  message. Instead we have to use a Zero-coded stream as a work around. Google advises not using protobug messages for large objects however, so perhaps a move away from this in future would be advisable.